Linelist data from an outbreak of Acute Jaundice Syndrome (AJS) in Chad 2016

Linelist data from a real outbreak of Hepatitis E virus (HEV) infection which occurred in the town of Am Timan, Chad, between October 2016 and April 2017.

Usage

AJS_AmTiman

Format

A data frame with 1477 rows and 122 columns:

Number

a numeric denoting HEV Count

HEVID

a numeric denoting HEV Case ID

Sex

a character denoting Sexe M=Male; F=Femelle

Age

a character denoting Age

Agegroup

a character denoting Groupe D'age 1=0 to 4; 2= 5 to 14; 3=15 to 44; 4==45

dateentry

a character denoting Initial Entry into Database (either Questionnaire or Medical Assessment)

epiquestdate

a character denoting Date de Questionnaire

team

a character denoting Equipe

chw

a character denoting AC

block

a character denoting Bloc

hh

a character denoting Menage

referredby

a character denoting Referer par: (1=Hôpital ; 2 = Superviseur Agent Communitaire)

quartier

a character denoting Quartier/Village

interviewee

a character denoting Statut interroge Patient ou Parent

hhpers

a character denoting Nombre de personnes dan le menage

hhmale

a character denoting Nombre masculin

hhfemale

a character denoting Nombre feminin

hhu5

a character denoting Nombre d'enfant < 5 ans

hpreferral

a character denoting Référé à l’hôpital par Agent Com? (O = Oui, N = Non)

hprefpos

a character denoting Si oui, le patient est-allé?

patientpresent

a character denoting Est-ce que le patient est présent ?

ptjaundice

a character denoting Est-ce que le patient a la jaunisse?

ptvomit

a character denoting Vomissements ?

mtmental

a character denoting Etat mental altéré ?

ptpregnant

a character denoting Enceinte?

ptu1

a character denoting Agée moins d’un an ?

referring

a character denoting 15. Allez-vous référer à l’hôpital?

otherajs

a character denoting Autre patient dans le foyer?

otherajsnum

a character denoting Si oui, combien?

borehole

a character denoting Forage?

comtap

a character denoting Bonne Fontaine?

river

a character denoting Riviere?

citywater

a character denoting Eau de la ville?

otherwater

a character denoting Autre

otherwatersp

a character denoting Si autre…

watersourcefreq

a character denoting Quel point utilisez-vous le plus souvent?

ceramicjar

a character denoting Jarre

jerrycan

a character denoting Jerrycan

oildrum

a character denoting Fut

reserother

a character denoting Autre?

reserothersp

a character denoting Si autre….

resertype

a character denoting 20. Type de stockage

chlorination

a character denoting 21. Ajouté chlore la derniere fois?

volumreser

a character denoting 22a. Combien de stockage?

handreser

a character denoting 22b. Entrer la main dans le stockage?

hweating

a character denoting 23. Laver les mains avant de manger?

hwwater

a character denoting 24a. Est-ce qu'il y a de l'eau?

hwsoap

a character denoting 24b. Est-ce qu'il y a du savon?

treatwater

a character denoting Traitez vous l'eau?

tptreatwater

a character denoting Type de traitement

tptreatwaterother

a character denoting Autre traitment

nottreatwhy

a character denoting Si non, pourquoi

defecriver

a character denoting Riviere?.1

defecair

a character denoting L'air libre?

defeclatrine

a character denoting Latrine?

numlatrine

a character denoting 26-Combin de personnes qui utilise le latrine?

dtmedical

a character denoting Date of Medical Assessment

hpid

a character denoting 1. Numero de Clinique/Hopital

outreachscreened

a character denoting Outreach screening? Y/N/U

evaljaundice

a character denoting 7. Deja Evalué pour la Jaunisse?

jaundiceid

a character denoting Si Oui, Numero de Cas?

referredbychw

a character denoting 8. Référé par l’Agent Communautaire?

jaundiceid2

a character denoting Si Oui, Numero de Cas?.1

medjaundice

a character denoting 9. Le patient a la jaunisse?

dtjaundice

a character denoting 10. Date de début de la jaunisse ?

medfever

a character denoting 11a. Fievre?

mednausea

a character denoting 11b. Nausées/Anorexie

medvomit

a character denoting 11c. Vomissements?

medepigastric

a character denoting 11d. Douleur epigastrique?

meditching

a character denoting 11e. Démangeaisons ?

medheadache

a character denoting 11f. Mal de tete?

medarthralgia

a character denoting 11g. Arthralgies?

meddiar

a character denoting 11h. Diarrhées

medbleeding

a character denoting 11i. Saignements?

medfev38

a character denoting 12. Fièvre = 38.0°C?

medmental

a character denoting 13. Etat Mental?

medother

a character denoting Other symptoms (free text)

medothhhajs

a character denoting 14. Autre avec jaunisse dans le menage?

medothhhajsnum

a character denoting 14a. Combien?

othercaseid

a character denoting 14b. Numero de cas

medpreg

a character denoting 15. Enceinte?

medpregtri

a character denoting 15a. Trimestre? 1/2/3

medpp

a character denoting 16. Post-Partum?

medppdays

a character denoting 16a. Si postpartum, combien de jours?

medppoutcome

a character denoting 16b. Resultat de la grossesse?

matdeathdt

a character denoting Si décès maternelle, date de décès

medppga

a character denoting Gestation weeks of delivery

medblood

a character denoting 17. Prelevements de sang?

medblooddt

a character denoting Si oui, date de prelevement

medmalrdt

a character denoting 18a. Malaria RDT

medhepb

a character denoting 18b. Hep B

medhepc

a character denoting 18c. Hep C

medhevrdt

a character denoting 18d. Hep E RDT

medhevelisa

a character denoting Hep E confiramtion ELISA MOH / MSF

hevrecent

a character denoting Confirmed 'recent' case based on ELISA enter 1 if IgM+

hevpcr

a character denoting Hep E confirmation PCR MOH/MSF

hevgenotest

a character denoting Genotype peformed? Y/N/unk

hevgenotype

a character denoting If yes, what type?

medhavelisa

a character denoting Hep A ELISA

medhbvelisa

a character denoting Hep Bc ELISA

medhcvelisa

a character denoting Hep C ELISA

arbovpcr

a character denoting ARBO PCR (YF, Rift, Chik, DEN, WN, ZIK)

yf

a character denoting Yellow Fever Elisa

zika

a character denoting Zika Elisa

dengue

a character denoting Dengue Elisa

wnv

a character denoting West Nile Elisa

rvf

a character denoting Rift Valley Elisa

chik

a character denoting Chikungunya Elisa

hospitalised

a character denoting 19. Hospitalised?

dthospitalisation

a character denoting Date de hopitalisation?

outcomehp

a character denoting 20. Resultat Clinique

dtdeath

a character denoting (Si décès la date de décès)

epiwajs

a character denoting EpiWeek Jaundice Onset

epiwhevelisa

a character denoting EpiWeek Confirmed (IgM+)

epiwoutreach

a character denoting EpiWeek FIRST Assessment (Community or Medical)

epiwmed

a character denoting EpiWeek MEDICAL Assessment

epiwepiquest

a character denoting EpiWeek Assessment QUESTIONNAIRE

epiwhp

a character denoting EpiWeek HOSPITALIZATION

epiwdeath

a character denoting EpiWeek DEATH

caseclass

a character denoting Suspected/Confirmed/Not A Case (based on HEV RDT)

Longitude

a numeric denoting Jittered GPS coordinates

Latitude

a numeric denoting Jittered GPS coordinates

Source

This data was collected as part of Doctors Without Borders (MSF) outbreak response and the the Ministry of Health (MoH) in Chad has approved the use of this data for training purposes.

The original dataset is AJS_AmTiman.xlsx.

This is version 1.0.

This dataset is English.

This dataset is licensed under GPL3

Details

This is a case-linelist which contains de-indentified individual level data on clinical, laboratory and epidemiological information. It can be used together with AJS_AmTiman_population, Blocksshape and Quartiersshape. Please note, that some data has been adapted in order to best achieve training objectives. Also, the GPS coordinates included in the dataset do not correspond to real cases identified during this outbreak. They have been generated exclusively for training purposes.

References

Spina A, Lenglet A, et al. (2017) A large outbreak of Hepatitis E virus genotype 1 infection in an urban setting in Chad likely linked to household level transmission factors, 2016-2017. PLoS ONE 12(11): e0188240 doi:10.1371/journal.pone.0188240