Shapefile of housing blocks from an outbreak of Acute Jaundice Syndrome (AJS) in Chad 2016

Shapefile data from a real outbreak of Hepatitis E virus (HEV) infection which occurred in the town of Am Timan, Chad, between October 2016 and April 2017.

Usage

Blocksshape

Format

A data frame with 650 rows and 3 columns:

variable1

explanation of a variable

variable2

more explanation of variables

Source

This data was collected as part of Doctors Without Borders (MSF) outbreak response and the the Ministry of Health (MoH) in Chad has approved the use of this data for training purposes.

The original dataset is 'Blocksshape.zip' (which contains the necessary Shapefile components).

This is version 1.0.

This dataset is English.

This dataset is licensed under GPL3

Details

This contains a shapefile for housing blocks which correspond with the population counts found in AJS_AmTiman_population and can also be used together with AJS_AmTiman.

References

Spina A, Lenglet A, et al. (2017) A large outbreak of Hepatitis E virus genotype 1 infection in an urban setting in Chad likely linked to household level transmission factors, 2016-2017. PLoS ONE 12(11): e0188240 doi:10.1371/journal.pone.0188240