Zip-code population data Fulton County (USA), 2020

Table of population size per zip code in Fulton County (USA), 2020

Usage

fulton_population

Format

An object of class data.frame with 49 rows and 2 columns.

Source

This population data was provided by Fulton County for training purposes.

This is version 1.0.

This dataset is English.

This dataset is licensed under CC by-NC-SA 4.0

Details

This is an aggregate dataset with population size per zip code in Fulton County (USA), 2020. It can be used together with fulton_data or fulton_data_clean, which are raw and clean versions of COVID-19 linelist data for Fulton County respectively.

Please note, that some data has been adapted in order to best achieve training objectives.