Fake data dictionary for a mortality survey

This is an example fake data dictionary for a mortality survey which was generated using epidict package functions.

Usage

mortality_survey_dict

Format

Two data frames, 'Mortality Survey' with 424 rows and 16 columns, and 'hh_member' with 2874 rows and 24 columns:

Source

This data was generated with functions from the epidict package which was created as part of the R4Epis project funded by the Doctors Without Borders (MSF) sappling nursery.

The original dataset is 'mortality_survey_dict.xlsx'

This is version 1.0.

This dataset is English.

This dataset is licensed under GPL3

Details

This is a fake data dictionary for the mortality_survey_data dataset. This data dictionary was originally in ODK excel format and using functions from epidict was transformed in to long format, so that it can be used for recoding using the matchmaker package. Please note, that this data is entirely fictional created based on the Doctors without Borders (MSF) ethical review board pre-approved survey dictionaries. They have been generated here exclusively for training purposes.