Test results for notifiable disease surveillance cases in Feveria in 2024 \(French\)

French version of the test results for notifiable disease surveillance cases notified in fictional country Feveria in 2024, for malaria, dengue, typhoid, yellow fever, and cholera.

Usage

multi_maladies_tests

Format

A dataframe with 1314 rows and 7 columns.

Source

Fictional data created by Applied Epi for training purposes

This is version 1.0.

This dataset is French.

This dataset is licensed under CC by-NC-SA 4.0

Details

This is the French version of a linelist with fictional test-level data from the imaginary country Feveria, with laboratory result information for suspected cases of malaria, dengue, typhoid, yellow fever, and cholera. The data is intended for use as part of the case study "Linking and analyzing notifiable disease surveillance data and laboratory data in R" alongside the French multi_maladies_notifications (notification data in Feveria).