Clean cholera surveillance data for Viraland, 17 June 2021

Clean surveillance data for a cholera outbreak in fictional country Viraland in 2021, with data as of 17 June.

Usage

viraland_cholera_20210617_linelist

Format

A dataframe with 833 rows and 38 columns.

Source

Fictional data created by Applied Epi for training purposes

This is a derivate (subset) of the clean cholera surveillance data for Viraland, 21 July 2021.

This dataset is English.

This dataset is licensed under CC by-NC-SA 4.0

Details

This is a linelist with fictional case-level data from the imaginary country Viraland, with epidemiological, lab, and clinical information for suspected cases of cholera. The data is intended for use alongside the intermediate qmd course and the intro to R course.