World “standard population” by sex

This is reference population counts which can be used together with country_demographics

Usage

world_standard_population_by_sex

Format

Dataframes

Source

This data is made available by Public Health Scotland

The original dataset is country_demographics.csv

This is version 1.0.

This dataset is English.

This dataset is licensed under OGL

Details

The World standard population is based on the populations of 46 countries and was developed in 1960. It is used to produce world standardised rates. It is a much ‘younger’ population compared to the population of Scotland (and other Western European countries). Also, like the European standard population, the World standard population has equal numbers of males and females within an age band. However, as the World standard population is a very ‘young’ population, it is generally less appropriate for Scottish data than the European standard population and is only be used for particular purposes. This file is split by five year age groups and sex